enable UChainGUI to change Udef for each unit, adding UdefListView
The name tags for the units are now DragSinks, which accept either a
string or symbol that is a valid name of an Udef, or an Udef object.
UdefListView provides a window with DragSources for all available
Udefs. So to change Udef; drag an Udef from the UdefListView to the
UChainGUI. All common parameters will remain set (i.e. if the current
and the new Udef both have a 'freq' arg, it will remain the same). The
UdefListView is organized by category, so I also added appropriate
categories to a number of UnitDefs.