adding URef
URef points to an object via an index relative to a parent. It can be
used to store pointers to objects in a file, without having to declare
variables. I'd like to use this for the modulation center (i.e. in
UModDest). A modulation center will be stored in a USession, and all
pointers are relative to that session. By storing the object they point
to as well we enable a possible resolving feature, which could find the
pointer again after an object has moved. This still needs some thought.