make drag and drop in UChainGUI work more logical
objects can now be dragged in between or on top of units. When dragged
in between the dragged object will always be inserted, not changing any
of the existing units. When dragging on top the unit will always be
replaced. When dragging a unit from one position in a chain on top of
another they will switch places.