fix for case sensitivity bug
as soon as we updated WFSCollider to 3.5 this can be replaced by
File.existsCaseSensitive (which isn't yet included in the current
WFSCollider source). For now, this basically does the same. Entering a
non-existing defName in a U now throws an error (which might also not
be what we want, but at least it doesn't start acting weird).