adding UChainGUI:singleWindow
off by default for now; single window mode makes all UChainGUI's appear
in the same window, effectively closing the previous one when a new one
is opened. This is to prevent lots of windows being open at the same
time. A button should be added to set this mode on/off, and perhaps it
should be on by default