implement lazy updating in UScoreView
this causes less cpu use during playback, allowing more complex scores
to still be playable. For scores with very many events however it is
still advised to either divide events over folders, or uncheck the
'update' button. Otherwise the graphics calculation may get in the way
of the audio timing.