add bounce function to UChain
bounce can bounce units in a chain to a single audio file. It will
place a \diskSoundFile unit with the recorded file in the chain. Bounce
only bounces private channels, if you want to bounce the whole unit
including output use writeAudioFile instead. The index is the index of
the last unit to be bounced. All units before this one are included.
The bounced file is a 32bits floating point .aiff file (multichannel if
needed).