fix UMarker pause function
This is a bit dirty, but it seems to work. It corrects for floating
point errors at stopping and starting. Without this sometimes objects
that are placed at the same start time as a pause marker fail to play
at unpause, especially when starting the score from somewhere in the
middle via GUI