first step to automatic random seed allocation
URandSeed can double as a Spec, and instances of the class could become
automatic seed generators, refreshing at each prepare call. They would
be the default object for a random seed arg. If the arg is an Integer
then the seed is fixed. Still some trickery is needed to move the seed
args to the end of the arg list, and a GUI to switch between the two