FreeUdef gets argSpecsFunc
a function to create the argspecs depending on the unit. Note that this
may cause issues with init. I.e. when an U is loaded from code and has
UMaps it doesn’t know its args yet when UMap:asUnitArgs is called,
which may cause the argSpecsFunc to fail. Please keep this in mind when
designing an argSpecsFunc